DEV : refactoring likelihoods, ParametricModel and cox
- adding other likelihoods to likelihoods module
- refactoring Cox
- simplifying ParametricModel, removing parameters names as it is not
used

TODO : remove fit from non parametric model (not used)
