+++ src/relife/likelihoods/__init__.py
@@ -1,3 +1,4 @@
+from ._cox_likelihood import init_cox_likelihood
 from ._lifetime_likelihood import LifetimeLikelihood
 
-__all__: list[str] = ["LifetimeLikelihood"]
+__all__: list[str] = ["LifetimeLikelihood", "init_cox_likelihood"]
